myaimentor-ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.6
openai==1.54.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-genai==1.73.1
